weather_map/pyproject.toml
[project]
name = "weather-map"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "logging>=0.4.9.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
